Added tests for auth, task, task_repository, user_repository and fixes after them

+++ internal/mocks/task_repository_mock.go
@@ -7,34 +7,54 @@ import (
 )
 
 type MockTaskRepository struct {
-	GetAllFunc  func(ctx context.Context) ([]model.Task, error)
-	GetByIDFunc func(ctx context.Context, id int) (*model.Task, error)
-	CreateFunc  func(ctx context.Context, task *model.Task) (sql.Result, error)
-	UpdateFunc  func(ctx context.Context, task *model.Task) error
-	DeleteFunc  func(ctx context.Context, id int) error
-	CountFunc   func(ctx context.Context) (int, error)
-}
-
-func (m *MockTaskRepository) GetAll(ctx context.Context) ([]model.Task, error) {
-	return m.GetAllFunc(ctx)
+	GetByIDFunc           func(ctx context.Context, id int) (*model.Task, error)
+	CreateFunc            func(ctx context.Context, task *model.Task) (sql.Result, error)
+	UpdateFunc            func(ctx context.Context, task *model.Task) error
+	DeleteFunc            func(ctx context.Context, id int) error
+	GetWithPaginationFunc func(ctx context.Context, limit, offset int) ([]model.Task, error)
+	CountFunc             func(ctx context.Context) (int, error)
 }
 
 func (m *MockTaskRepository) GetByID(ctx context.Context, id int) (*model.Task, error) {
-	return m.GetByIDFunc(ctx, id)
+	if m.GetByIDFunc != nil {
+		return m.GetByIDFunc(ctx, id)
+	}
+
+	return nil, sql.ErrNoRows
 }
 
 func (m *MockTaskRepository) Create(ctx context.Context, task *model.Task) (sql.Result, error) {
-	return m.CreateFunc(ctx, task)
+	if m.CreateFunc != nil {
+		return m.CreateFunc(ctx, task)
+	}
+
+	return nil, nil
 }
 
 func (m *MockTaskRepository) Update(ctx context.Context, task *model.Task) error {
-	return m.UpdateFunc(ctx, task)
+	if m.UpdateFunc != nil {
+		return m.UpdateFunc(ctx, task)
+	}
+	return nil
 }
 
 func (m *MockTaskRepository) Delete(ctx context.Context, id int) error {
-	return m.DeleteFunc(ctx, id)
+	if m.DeleteFunc != nil {
+		return m.DeleteFunc(ctx, id)
+	}
+	return nil
+}
+
+func (m *MockTaskRepository) GetWithPagination(ctx context.Context, limit, offset int) ([]model.Task, error) {
+	if m.GetWithPaginationFunc != nil {
+		return m.GetWithPaginationFunc(ctx, limit, offset)
+	}
+	return []model.Task{}, nil
 }
 
 func (m *MockTaskRepository) Count(ctx context.Context) (int, error) {
-	return m.CountFunc(ctx)
+	if m.CountFunc != nil {
+		return m.CountFunc(ctx)
+	}
+	return 0, nil
 }
